+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/spring/core/setup/SpringProxyJobClassNameProvider.java
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.spring.namespace.setup;
+package org.apache.shardingsphere.elasticjob.lite.spring.core.setup;
 
+import lombok.extern.slf4j.Slf4j;
 import org.apache.shardingsphere.elasticjob.api.ElasticJob;
 import org.apache.shardingsphere.elasticjob.lite.internal.setup.JobClassNameProvider;
-import org.apache.shardingsphere.elasticjob.lite.spring.namespace.job.util.AopTargetUtils;
+import org.apache.shardingsphere.elasticjob.lite.spring.core.util.AopTargetUtils;
 import org.springframework.aop.support.AopUtils;
 
 /**
@@ -29,8 +30,13 @@ import org.springframework.aop.support.AopUtils;
  * Consider the proxy object that generated by cglib or jdk dynamic proxy.
  * </p>
  */
+@Slf4j
 public final class SpringProxyJobClassNameProvider implements JobClassNameProvider {
-    
+
+    public SpringProxyJobClassNameProvider() {
+        log.info("create SpringProxyJobClassNameProvider");
+    }
+
     @Override
     public String getJobClassName(final ElasticJob elasticJob) {
         return AopUtils.isAopProxy(elasticJob) ? AopTargetUtils.getTarget(elasticJob).getClass().getName() : elasticJob.getClass().getName();
