+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.internal.setup.JobClassNameProvider
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.spring.namespace.setup.SpringProxyJobClassNameProvider
+org.apache.shardingsphere.elasticjob.lite.spring.core.setup.SpringProxyJobClassNameProvider
